node: Split source file into a full Python package
